code/API/requirements.txt
streamlit
Pillow
pydeck
google-cloud-bigquery
googlemaps
statsmodels
matplotlib
folium
streamlit-folium
pytyped-metrics
streamlit-extras
streamlit-elements
streamlit-option-menu